+++ LEDGeometry/MonoColorEffect.cpp
@@ -1,6 +1,7 @@
 #include "MonoColorEffect.h"
 
 #include <FastLED.h>
+#include <stdint.h>
 
 #include "ColorScheduler.h"
 #include "LEDCurve.h"
@@ -9,7 +10,7 @@
 namespace LEDGeometry {
 void MonoColorEffect::update(LEDCurve* led_curve) {
     CRGB color = led_curve->color_scheduler->next_color();
-    for (int i = 0; i < led_curve->shape->n_points(); i++) {
+    for (uint8_t i = 0; i < led_curve->shape->n_points(); i++) {
         led_curve->leds[i] = color;
     }
 }
