+++ LEDGeometry/PulseEffect.cpp
@@ -10,13 +10,13 @@ namespace LEDGeometry {
 
 PulseEffect::PulseEffect() : PulseEffect(30) {}
 
-PulseEffect::PulseEffect(int beats_per_minute)
+PulseEffect::PulseEffect(uint8_t beats_per_minute)
     : beats_per_minute(beats_per_minute) {}
 
 void PulseEffect::update(LEDCurve* led_curve) {
     CRGB color = led_curve->color_scheduler->next_color();
-    color.fadeLightBy(beatsin8((uint8_t)beats_per_minute));
-    for (int i = 0; i < led_curve->shape->n_points(); i++) {
+    color.fadeLightBy(beatsin8(beats_per_minute));
+    for (uint8_t i = 0; i < led_curve->shape->n_points(); i++) {
         led_curve->leds[i] = color;
     }
 }
