+++ LEDGeometry/PulseEffect.h
@@ -1,4 +1,6 @@
 #pragma once
+#include <stdint.h>
+
 #include "LightEffect.h"
 
 namespace LEDGeometry {
@@ -7,12 +9,12 @@ class LEDCurve;
 class PulseEffect : public LightEffect {
    public:
     PulseEffect();
-    PulseEffect(int beats_per_minute);
+    PulseEffect(uint8_t beats_per_minute);
 
    protected:
     void update(LEDCurve* led_curve);
 
    private:
-    int beats_per_minute;
+    uint8_t beats_per_minute;
 };
 }  // namespace LEDGeometry
