+++ LEDGeometry/SignalTransmissionEffect.cpp
@@ -7,15 +7,15 @@
 #include "Shape.h"
 
 namespace LEDGeometry {
-SignalTransmissionEffect::SignalTransmissionEffect(int fade_factor)
+SignalTransmissionEffect::SignalTransmissionEffect(uint8_t fade_factor)
     : fade_factor(fade_factor), pos(0) {}
 
 SignalTransmissionEffect::SignalTransmissionEffect()
     : SignalTransmissionEffect(16) {}
 
 void SignalTransmissionEffect::update(LEDCurve* led_curve) {
-    int n_points = led_curve->shape->n_points();
-    for (int i = 0; i < n_points; i++) {
+    uint8_t n_points = led_curve->shape->n_points();
+    for (uint8_t i = 0; i < n_points; i++) {
         led_curve->leds[i].fadeLightBy(fade_factor);
     }
     ++pos;
