+++ LEDGeometry/SignalTransmissionEffect.h
@@ -1,4 +1,6 @@
 #pragma once
+#include <stdint.h>
+
 #include "LightEffect.h"
 
 namespace LEDGeometry {
@@ -7,13 +9,13 @@ class LEDCurve;
 class SignalTransmissionEffect : public LightEffect {
    public:
     SignalTransmissionEffect();
-    SignalTransmissionEffect(int fade_factor);
+    SignalTransmissionEffect(uint8_t fade_factor);
 
    protected:
     void update(LEDCurve* led_curve);
 
    private:
-    int fade_factor;
-    int pos;
+    uint8_t fade_factor;
+    uint8_t pos;
 };
 }  // namespace LEDGeometry
